+++ src/ui/components/AvatarComponent.jsx
