+++ src/ui/components/AvatarComponent.module.css
